requirements.txt
anyio==3.7.0
click==8.1.3
colorama==0.4.6
fastapi==0.98.0
flake8==6.0.0
flake8-isort==6.0.0
h11==0.14.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
pycodestyle==2.10.0
pydantic==1.10.9
pyflakes==3.0.1
pyserial==3.5
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
